+++ docs/source/conf.py
@@ -10,7 +10,7 @@ import sys
 
 sys.path.insert(0, os.path.abspath(".."))
 
-__version__ = "2025.06.08.2"
+__version__ = "2025.06.23.1"
 
 project = "django-blocknote"
 copyright = "2025, Ryan Sevelj & Mark Sevelj"
@@ -27,11 +27,22 @@ extensions = [
     "sphinx_copybutton",
     "sphinx_inline_tabs",
     "sphinx.ext.todo",
+    "sphinxcontrib.mermaid",
 ]
+mermaid_output_format = "raw"  # or 'png'
+mermaid_params = ["--theme", "default"]
+
 myst_enable_extensions = [
     "colon_fence",
-    "mermaid",
+    "html_admonition",
 ]
+
+source_suffix = {
+    ".rst": "restructuredtext",
+    ".md": "markdown",
+}
+
+
 """
 The colon fence syntax is particularly useful because: